Add README only if not exists

+++ create_app.sh
@@ -33,19 +33,19 @@ pm_format() {
     case $package_manager in
         "npm")
             echo -e "$RED$package_manager"
-            ;;
+        ;;
         "pipenv")
             echo -e "$YELLOW$package_manager"
-            ;;
+        ;;
         "yarn")
             echo -e "$CYAN$package_manager"
-            ;;
+        ;;
         "intelliJ")
             echo -e "$MAGENTA$package_manager"
-            ;;
+        ;;
         *)
             echo "$package_manager"
-            ;;
+        ;;
     esac
 }
 
@@ -55,28 +55,28 @@ template_format() {
     case $template in
         "web")
             echo -e "$YELLOW$template"
-            ;;
+        ;;
         "javascript" | "js")
             template="JavaScript"
             echo -e "$YELLOW$template"
-            ;;
+        ;;
         "py" | "python")
             template="Python"
             echo -e "$YELLOW$template"
-            ;;
+        ;;
         "typescript" | "ts")
             template="TypeScript"
             echo -e "$BLUE$template"
-            ;;
+        ;;
         "go")
             echo -e "$BLUE$template"
-            ;;
+        ;;
         "java")
             echo -e "$RED$template"
-            ;;
+        ;;
         *)
             echo "$template"
-            ;;
+        ;;
     esac
 }
 
@@ -85,9 +85,9 @@ is_template_valid() {
 
     if [ "$template" == "web" ]; then
         return 0
-    elif [ "$template" == "py" ] || [ "$template" == "python" ]; then
+        elif [ "$template" == "py" ] || [ "$template" == "python" ]; then
         return 0
-    elif [ "$template" == "java" ]; then
+        elif [ "$template" == "java" ]; then
         return 0
     else
         echo -e "$RED"
@@ -190,29 +190,29 @@ config_args() {
     case $arg in
         "--update" | "-u")
             update_app
-            ;;
+        ;;
         "--help" | "-h")
             prompt_help
-            ;;
+        ;;
         "--templates" | "-t")
             templates
-            ;;
+        ;;
         "--set-package-manager" | "--set-pm" | "-sp")
             change_package_manager "$opt"
-            ;;
+        ;;
         "--set-template" | "-st")
             change_template "$opt"
-            ;;
+        ;;
         "--version" | "-v")
-            ;;
+        ;;
         "--defaults" | "-d")
             defaults
-            ;;
+        ;;
         *)
             echo "Unknown option \"$arg\""
             echo "    Try: create-app --help"
             exit 1
-            ;;
+        ;;
     esac
     exit 0
 
@@ -247,7 +247,9 @@ template_setup() {
         git clone -b "$template" -q "$repo_url" "$project_name"
         } && {
         cd "$project_name"
-        echo "# $project_name" > README.md
+        if [ ! -f README.md ]; then
+            echo "# $project_name" > README.md
+        fi
         rm -rf .git/
         git init
         git checkout -b main
